mmap good

+++ kernel/proc.h
@@ -99,17 +99,17 @@ struct vma
 {
     /* vma info */
     int occupied;   // denote whether occupied(1) or not(0)
-    int start_addr; // starting address of vma
-    int end_addr; // ending address of vma
+    uint64 start_addr; // starting address of vma
+    uint64 end_addr; // ending address of vma
     // struct file * p_file;
 
     /* fetched arguments */
     uint64 addr; // arg0, kernel chooses virtual address
-    int length;  // arg1, indicates how many bytes to map
+    uint64 length;  // arg1, indicates how many bytes to map
     int prot;    // arg2, indicates the mapped memory mode
     int flags;   // arg3, the flags of mapping
     int fd;      // arg4, file description
-    int offset;  // arg5, the starting offset in the file
+    uint64 offset;  // arg5, the starting offset in the file
     struct file *pf;
 };
